[refactoring] lottonumber로 이름 변경
